+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.8] - 2020-10-26
+
+* Add WithoutTraits filter in semantic query editor
+* Fix light UI theme
+
 ## [0.2.7] - 2020-10-26
 
 * Update to Entities 0.16